:art: 타입 파일 병합
- DataTypes.ts와 type.ts를 한 파일로 병합했습니다.
