+++ frontend/src/type.ts
@@ -1,3 +1,52 @@
+export type StatusResponseExpired = {
+  intra_id?: string;
+  floor?: number;
+  cabinet_num?: number;
+  expire_time?: string;
+};
+
+export type StatusResponseExpiredData = StatusResponseDisabled[];
+
+export type StatusResponseDisabled = {
+  intra_id?: string;
+  floor?: number;
+  cabinet_num?: number;
+  expire_time?: string;
+};
+
+export type StatusResponseDisabledData = StatusResponseDisabled[];
+
+export type SearchResponseFromLent = {
+  activation?: number;
+  cabinet_id?: number;
+  cabinet_num?: number;
+  expire_time?: string;
+  floor?: number;
+  intra_id?: string;
+  lent_time?: string;
+  location?: string;
+  section?: string;
+  lent_id?: number;
+};
+
+export type SearchResponseFromLentLog = {
+  activation?: number;
+  cabinet_id?: number;
+  cabinet_num?: number;
+  return_time?: string;
+  floor?: number;
+  intra_id?: string;
+  lent_time?: string;
+  location?: string;
+  section?: string;
+  lent_id?: number;
+};
+
+export type SearchResponseData = {
+  resultFromLent?: SearchResponseFromLent[];
+  resultFromLentLog?: SearchResponseFromLentLog[];
+};
+
 export type FloorStateData = {
   disabled: number;
   floor: string;
